Handle int/long changes in Python 3.
Python 3 has no separate "long" type; instead, all ints are arbitrary-sized
"long" integers.  Slightly confusingly, they're called int in Python but
PyLongObject etc. in C.

At the C level, the shortcut for ints in _size_of no longer applies in
Python 3, and we need to make some other slight adjustments to handle longs
in a few places.  Conveniently, Python 2's PyLong_AsLong works with both
ints and longs, although this isn't true for all PyLong_* functions and
macros.

At the Python level, use the popular "six" library to help paper over the
differences.  I intend to make more use of this in future Python 3 porting
branches, so while it could easily be written out longhand in this case it
will pay its way overall.  We need to be a little careful in one test that
used long literals (ending with "L"), which are no longer valid in Python 3;
while the test itself can simply be skipped in Python 3 since there's a
similar version for ints immediately beforehand, we still need to make sure
to use a construction that isn't a syntax error.