Update meliae.tests.test__scanner to handle Python 3's bytes/text changes.
Porting these tests first makes it easier to verify that later changes don't
introduce accidental behaviour changes.

I've adopted the terminology of "bytes" and "text" rather than "string" and
"unicode" to make it clearer which set of semantics is needed in each case,
regardless of Python version.  The terms "string" and "str" are now
generally reserved for cases where we actually mean the native str type,
whatever that is; for instance, the __name__ of a class is always a str, so
bytes on Python 2 and text on Python 3.

_bytes_to_json needs a few minor contortions to cope with the differences
between iterating over bytes objects in Python 2 and 3; six.iterbytes deals
with most of this.  _text_to_json requires some adjustments because
bytes.encode doesn't exist any more, and neither do ur'...' literals.

TestPyDumpJSONObj.assertDumpBytes accepts either bytes or text as its
expected argument.  This is mainly to make it easy to support Python 3.4,
where bytes objects don't support the % operator (this was restored in 3.5).