Skip dumping large longs that would require allocation.
Python long objects (spelled "int" at the Python level in Python 3) are
arbitrary-precision integers, so simply converting them to C longs using
PyLong_AsLong might overflow and raise OverflowError.  (I ran into such a
case in the Launchpad test suite; I didn't track down exactly what it was,
but a Python long somewhere in memory that doesn't fit into a C long isn't
all that implausible.)

Converting to long long instead (supported as of Python 2.7) makes this
situation less likely, but it's still possible, and I don't know of a way to
accurately format an arbitrary-precision integer to decimal without
allocating memory, which we try to avoid in the inner parts of the scanner.
Therefore, if the conversion overflows, just omit the value from the dump.
This is presumably reasonably tolerable since longs weren't dumped at all
before meliae 0.5.